Move request properties and methods to middleware.py
define.py has too much stuff in it already.
