+++ weasyl/test/conftest.py
@@ -10,7 +10,7 @@ config._in_test = True  # noqa
 
 from libweasyl.configuration import configure_libweasyl
 from libweasyl.models.tables import metadata
-from weasyl import cache, define, emailer, macro, media
+from weasyl import cache, define, emailer, macro, media, middleware
 cache.region.configure('dogpile.cache.memory')
 define.metric = lambda *a, **kw: None
 
@@ -35,10 +35,10 @@ def setupdb(request):
 @pytest.fixture(autouse=True)
 def setup_request_environment(request):
     pyramid_request = pyramid.testing.DummyRequest()
-    pyramid_request.set_property(define.pg_connection_request_property, name='pg_connection', reify=True)
-    pyramid_request.set_property(define.userid_request_property, name='userid', reify=True)
-    pyramid_request.log_exc = define.log_exc_request_method
-    pyramid_request.web_input = define.web_input_request_method
+    pyramid_request.set_property(middleware.pg_connection_request_property, name='pg_connection', reify=True)
+    pyramid_request.set_property(middleware.userid_request_property, name='userid', reify=True)
+    pyramid_request.log_exc = middleware.log_exc_request_method
+    pyramid_request.web_input = middleware.web_input_request_method
     pyramid_request.environ['HTTP_X_FORWARDED_FOR'] = '127.0.0.1'
     pyramid_request.client_addr = '127.0.0.1'
     pyramid.testing.setUp(request=pyramid_request)
